Extract UIView helpers to a separate extension
