+++ Kickstarter-iOS/Views/Controllers/PaymentMethodsViewController.swift
@@ -205,19 +205,6 @@ extension PaymentMethodsViewController: AddNewCardViewControllerDelegate {
   }
 }
 
-private extension UIView {
-  func constrainEdges(to view: UIView) {
-    self.translatesAutoresizingMaskIntoConstraints = false
-
-    NSLayoutConstraint.activate([
-      self.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-      self.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-      self.topAnchor.constraint(equalTo: view.topAnchor),
-      self.bottomAnchor.constraint(equalTo: view.bottomAnchor)
-      ])
-  }
-}
-
 private extension UITableView {
   func ksr_sizeHeaderFooterViewsToFit() {
     let keyPaths: [ReferenceWritableKeyPath<UITableView, UIView?>] = [
